Add Hawk Results doc and change batch size in token selector

+++ home/script_labeler.py
@@ -32,7 +32,7 @@ class ScriptLabeler:
             init_samples = selector_field['token']['initial_samples']
             num_scouts = len(self.configuration['scouts'])
             self.total_init_samples = int(init_samples * int(num_scouts))
-            self._label_time = int(selector_field['token']['label_time'])
+            self._label_time = float(selector_field['token']['label_time'])
         ##########
 
         if label_mode == "classify":
